cast literal int to int64 to fix integration tests.

+++ insert_test.go
@@ -59,7 +59,7 @@ func TestInsertRecordsToSql(t *testing.T) {
 	sql, args := s.InsertInto("a").Columns("something_id", "user_id", "other").Record(objs[0]).Record(objs[1]).ToSql()
 
 	assert.Equal(t, sql, "INSERT INTO a (`something_id`,`user_id`,`other`) VALUES (?,?,?),(?,?,?)")
-	assert.Equal(t, args, []interface{}{1, 88, false, 2, 99, true})
+	assert.Equal(t, args, []interface{}{1, int64(88), false, 2, int64(99), true})
 }
 
 func TestInsertKeywordColumnName(t *testing.T) {
@@ -70,7 +70,7 @@ func TestInsertKeywordColumnName(t *testing.T) {
 
 	rowsAff, err := res.RowsAffected()
 	assert.NoError(t, err)
-	assert.Equal(t, rowsAff, 1)
+	assert.Equal(t, rowsAff, int64(1))
 }
 
 func TestInsertReal(t *testing.T) {
@@ -101,7 +101,7 @@ func validateInsertingBarack(t *testing.T, s *Session, res sql.Result, err error
 	assert.NoError(t, err)
 
 	assert.True(t, id > 0)
-	assert.Equal(t, rowsAff, 1)
+	assert.Equal(t, rowsAff, int64(1))
 
 	var person dbrPerson
 	err = s.Select("*").From("dbr_people").Where("id = ?", id).LoadStruct(&person)
